[WIP] Encoding YFC100M dataset.
The notebook does not work. We need to filter out entries that do not
have a corresponding image, but I'm having trouble doing it. Pandas
struggles with memory (in my system, with 96 GB). Datasets is great but
gets confused sometimes when the map function returns no entries.

Note also that I disabled the encoding function itself (it's returning a
random list of tokens), because flax requires system-wide availability of
cudNN, apparently.